// src/domain/enums/AppointmentStatus.ts
export enum AppointmentStatus {
  Active = 'Activa',
  SecretaryReview = 'En revision de secretaria',
  WaitingVitalSigns = 'En espera de signos vitales',
  ReadyForAppointment = 'Lista para consulta',
  Rescheduled = 'Reagendada',
  Started = 'Iniciada',
  Completed = 'Completada',
  Cancelled = 'Cancelada',
}